Iron Zombie purging, world gen changes
I also attempted a broken fix to the super slime, however all it does is
spam quite a few errors. Iron zombie purging exists due to a change in
the Iron Zombie's data watcher in 1.8 that would make all 1.7 Iron
Zombies vomit errors to the console. However, it too, is broken.